Minor corrections for VESPA/Delft

+++ System/attachCompInc/FixedComp.h
@@ -141,7 +141,7 @@ class FixedComp
   /// Access beamOrigin
   virtual const Geometry::Vec3D& getBeamOrigin() const { return beamOrigin; }  
 
-  virtual int getExitWindow(const size_t,std::vector<int>&) const;
+  virtual int getExitWindow(const long int,std::vector<int>&) const;
   virtual int getMasterSurf(const size_t) const;
 			 
   void copyLinkObjects(const FixedComp&);
